requirements.txt
celery
redis 
requests
python-dotenv 
